Move topology to integrations
