+++ test/src/topology_test.cpp
@@ -4,7 +4,7 @@
 #include "fff.h"
 
 extern "C" {
-#include "messages/topology.h"
+#include "topology.h"
 #include "mock_bcmp.h"
 #include "mock_bm_os.h"
 #include "mock_device.h"
